Revert "Dynamic stock retrieval"
