+++ .idea/misc.xml
@@ -1,8 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
-  <component name="Black">
-    <option name="sdkName" value="Python 3.11 (investment-calculator)" />
-  </component>
   <component name="ProjectRootManager" version="2" languageLevel="JDK_20">
     <output url="file://$PROJECT_DIR$/out" />
   </component>
